Quill-270 Split the web widget theme editor into a shell, inspector and preview stage
