+++ src/Raven.Quill.Web/src/pages/apps/channels/theme-editor/theme-editor-fields.ts
@@ -1,6 +1,13 @@
-import type { FieldPath } from "react-hook-form";
+import type { Control, FieldPath } from "react-hook-form";
 import type { PreviewAppearance } from "@/pages/apps/channels/web-widget-theme-preview";
 import type { WidgetThemeFormData } from "@/pages/apps/channels/web-widget-theme-schema";
+
+/** Shared by every theme editor section so the inspector can render them uniformly. */
+export type ThemeSectionProps = {
+    control: Control<WidgetThemeFormData>;
+    isSaving: boolean;
+    onReset: (paths: readonly FieldPath<WidgetThemeFormData>[]) => void;
+};
 
 export const SCHEME_COLOR_FIELDS = {
     Light: ["lightButtonColor", "lightMessageColor", "lightBackgroundColor"],
